requirements.txt
requests
python-dotenv
notion-client
openai
pytrends
